Move lazy var to separate file
